Refactor DB init
Slightly changed how the database is initialized. The way it works right now is straightforward: in the main window, call `DatabaseManager::loadDatabase()`. If not error is returned, go ahead, else halt execution. Instances of `DatabaseManager` do not have an `ok()` method anymore and queries will simply fail if the DB is not initialized.